backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.115.4
uvicorn[standard]==0.27.0
pydantic==2.10.2
pydantic-settings==2.6.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.7

# Environment
python-dotenv==1.0.1

# Code Quality
ruff==0.8.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0